refactor(plotters): drop legacy
