+++ oceanarray/plotters/__init__.py
@@ -23,10 +23,9 @@ Available now (pre-OdB):
 
 Legacy functions (from plotter.py — available via backward-compat shim):
 
-  plot_qartod_summary, plot_climatology, scatter_profile_vs_PRES,
-  pcolor_timeseries_by_depth, plot_timeseries_by_depth, plot_trim_windows,
+  plot_qartod_summary, plot_trim_windows,
   plot_microcat_raw, plot_aquadopp_raw, plot_microcat,
-  show_variables, show_attributes, plot_mooring_timeseries, plot_grid
+  plot_mooring_timeseries
 
 As each legacy function is migrated into the appropriate sub-module, remove
 it from plotter.py and from the shim below.  See the migration checklist at
@@ -58,19 +57,17 @@ from oceanarray.plotters._animation import animate_hodograph  # noqa: F401
 # ---------------------------------------------------------------------------
 from oceanarray.plotter import (  # noqa: F401
     plot_qartod_summary,
-    plot_climatology,
-    scatter_profile_vs_PRES,
-    pcolor_timeseries_by_depth,
-    plot_timeseries_by_depth,
     plot_trim_windows,
     plot_microcat_raw,
     plot_aquadopp_raw,
     plot_microcat,
-    show_variables,
-    show_attributes,
     plot_mooring_timeseries,
-    plot_grid,
 )
+
+# Relocated out of plotter.py (2026-08-06):
+#   show_variables/show_attributes -> oceanarray.inspect.vars / .attrs
+#   plot_climatology               -> oceanarray.tools.rapid_interp.plot_climatology
+#   plot_grid._panel               -> plotters._primitives.pcolormesh_panel (plot_grid deleted)
 
 __all__ = [
     # New (pre-OdB)
@@ -82,16 +79,9 @@ __all__ = [
     "animate_hodograph",
     # Legacy shim
     "plot_qartod_summary",
-    "plot_climatology",
-    "scatter_profile_vs_PRES",
-    "pcolor_timeseries_by_depth",
-    "plot_timeseries_by_depth",
     "plot_trim_windows",
     "plot_microcat_raw",
     "plot_aquadopp_raw",
     "plot_microcat",
-    "show_variables",
-    "show_attributes",
     "plot_mooring_timeseries",
-    "plot_grid",
 ]
